+++ modules/angular2/src/render/dom/events/hammer_common.js
