+++ modules/angular2/src/render/dom/events/hammer_gestures.dart
@@ -2,7 +2,7 @@ library angular.events;
 
 import 'dart:html';
 import './hammer_common.dart';
-import '../../facade/lang.dart' show BaseException;
+import 'package:angular2/src/facade/lang.dart' show BaseException;
 
 import 'dart:js' as js;
