+++ modules/angular2/src/render/dom/events/hammer_gestures.es6
