+++ modules/angular2/src/services/url_resolver.js
