+++ modules/angular2/test/core/compiler/style_inliner_spec.js
@@ -17,7 +17,7 @@ import {isBlank} from 'angular2/src/facade/lang';
 import {Promise, PromiseWrapper} from 'angular2/src/facade/async';
 import {Map, MapWrapper} from 'angular2/src/facade/collection';
 
-import {XHR} from 'angular2/src/core/compiler/xhr/xhr';
+import {XHR} from 'angular2/src/services/xhr';
 
 import {bind} from 'angular2/di';
