+++ modules/angular2/test/core/compiler/style_url_resolver_spec.js
@@ -1,7 +1,7 @@
 import {describe, it, expect, beforeEach, ddescribe, iit, xit, el} from 'angular2/test_lib';
 import {StyleUrlResolver} from 'angular2/src/core/compiler/style_url_resolver';
 
-import {UrlResolver} from 'angular2/src/core/compiler/url_resolver';
+import {UrlResolver} from 'angular2/src/services/url_resolver';
 
 export function main() {
   describe('StyleUrlResolver', () => {
