+++ modules/benchmarks/src/largetable/largetable_benchmark.js
@@ -12,7 +12,7 @@ import {ShadowDomStrategy, NativeShadowDomStrategy, EmulatedUnscopedShadowDomStr
 import {Content} from 'angular2/src/core/compiler/shadow_dom_emulation/content_tag';
 import {DestinationLightDom} from 'angular2/src/core/compiler/shadow_dom_emulation/light_dom';
 import {LifeCycle} from 'angular2/src/core/life_cycle/life_cycle';
-import {UrlResolver} from 'angular2/src/core/compiler/url_resolver';
+import {UrlResolver} from 'angular2/src/services/url_resolver';
 import {StyleUrlResolver} from 'angular2/src/core/compiler/style_url_resolver';
 import {ComponentUrlMapper} from 'angular2/src/core/compiler/component_url_mapper';
 import {StyleInliner} from 'angular2/src/core/compiler/style_inliner';
@@ -26,13 +26,13 @@ import {isPresent, BaseException} from 'angular2/src/facade/lang';
 import {window, document, gc} from 'angular2/src/facade/browser';
 import {getIntParameter, getStringParameter, bindAction} from 'angular2/src/test_lib/benchmark_util';
 
-import {XHR} from 'angular2/src/core/compiler/xhr/xhr';
-import {XHRImpl} from 'angular2/src/core/compiler/xhr/xhr_impl';
+import {XHR} from 'angular2/src/services/xhr';
+import {XHRImpl} from 'angular2/src/services/xhr_impl';
 
 import {If, For, Switch, SwitchWhen, SwitchDefault} from 'angular2/directives';
 import {BrowserDomAdapter} from 'angular2/src/dom/browser_adapter';
 
-import {EventManager} from 'angular2/src/core/events/event_manager';
+import {EventManager} from 'angular2/src/render/dom/events/event_manager';
 
 import {ListWrapper} from 'angular2/src/facade/collection';
 import {Parent} from 'angular2/src/core/annotations/visibility';
