+++ modules/benchmarks/src/naive_infinite_scroll/index.js
@@ -17,9 +17,9 @@ import {DestinationLightDom} from 'angular2/src/core/compiler/shadow_dom_emulati
 import {TemplateLoader} from 'angular2/src/core/compiler/template_loader';
 import {TemplateResolver} from 'angular2/src/core/compiler/template_resolver';
 import {LifeCycle} from 'angular2/src/core/life_cycle/life_cycle';
-import {XHR} from 'angular2/src/core/compiler/xhr/xhr';
-import {XHRImpl} from 'angular2/src/core/compiler/xhr/xhr_impl';
-import {UrlResolver} from 'angular2/src/core/compiler/url_resolver';
+import {XHR} from 'angular2/src/services/xhr';
+import {XHRImpl} from 'angular2/src/services/xhr_impl';
+import {UrlResolver} from 'angular2/src/services/url_resolver';
 import {StyleUrlResolver} from 'angular2/src/core/compiler/style_url_resolver';
 import {ComponentUrlMapper} from 'angular2/src/core/compiler/component_url_mapper';
 import {StyleInliner} from 'angular2/src/core/compiler/style_inliner';
@@ -36,7 +36,7 @@ import {CompanyNameComponent, OpportunityNameComponent, OfferingNameComponent,
     setupReflectorForCells}
         from './cells';
 
-import {EventManager} from 'angular2/src/core/events/event_manager';
+import {EventManager} from 'angular2/src/render/dom/events/event_manager';
 
 export function main() {
   setupReflector();
